+++ src/components/app/userActionFormEmailCongressperson/index.tsx
@@ -39,7 +39,7 @@ import { PageTitle } from '@/components/ui/pageTitleText'
 import { ScrollArea } from '@/components/ui/scroll-area'
 import { Textarea } from '@/components/ui/textarea'
 import { useCountryCode } from '@/hooks/useCountryCode'
-import { useGetDTSIPeopleFromAddress } from '@/hooks/useGetDTSIPeopleFromAddress'
+import { useGetDTSIPeopleFromAddress } from '@/hooks/useGetDTSIPeopleFromUSAddress'
 import { useIntlUrls } from '@/hooks/useIntlUrls'
 import { useIsDesktop } from '@/hooks/useIsDesktop'
 import { convertAddressToAnalyticsProperties } from '@/utils/shared/sharedAnalytics'
